[DPM-647] fix duplicate events
