+++ go.mod
@@ -10,6 +10,7 @@ require (
 	github.com/google/uuid v1.1.1
 	github.com/gorilla/mux v1.7.4
 	github.com/gorilla/websocket v1.4.2
+	github.com/jackc/pgconn v1.4.0
 	github.com/jackc/pgtype v1.2.0
 	github.com/jackc/pgx/v4 v4.5.0
 	github.com/kelseyhightower/envconfig v1.4.0
